Convert some of the functions to async/await

+++ server/server.js
@@ -27,47 +27,39 @@ app.use(bodyParser.json());
  ██████  ███████ ███████ ██   ██     ██   ██  ██████   ██████     ██    ███████ ███████
 ******************************************************************************/
 
-app.post('/users', (req, res) => {
-  var user = new User({
-    email: req.body.email,
-    password: req.body.password
-  });
-
-  user.save().then(() => {
-    return user.generateAuthToken();
-  }).then((token) => {
+app.post('/users', async (req, res) => {
+  try {
+    const user = new User( {email: req.body.email, password: req.body.password} );
+    await user.save();
+    const token = await user.generateAuthToken();
     res.header('x-auth', token).send(user);
-  }).catch((error) => {
+  } catch (err) {
     res.status(400).send();
-  });
+  }
 });
-
 
 app.get('/users/me', authenticate, (req, res) => {
   res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-  var userReq = new User({
-    email: req.body.email,
-    password: req.body.password
-  });
-
-  User.findByCredentials(userReq.email, userReq.password).then((user) => {
-    return user.generateAuthToken().then((token) => {
-      res.header('x-auth', token).send(user);
-    });
-  }).catch((error) => {
+app.post('/users/login', async (req, res) => {
+  try {
+    const userReq = new User( {email: req.body.email, password: req.body.password} );
+    const user = await User.findByCredentials(userReq.email, userReq.password);
+    const token = await user.generateAuthToken();
+    res.header('x-auth', token).send(user);
+  } catch (err) {
     res.status(400).send();
-  });
+  }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-  req.user.removeToken(req.token).then(() => {
+app.delete('/users/me/token', authenticate, async (req, res) => {
+  try {
+    await req.user.removeToken(req.token);
     res.status(200).send();
-  }, () => {
+  } catch (err) {
     res.status(400).send();
-  });
+  }
 });
 
 
@@ -127,24 +119,22 @@ app.get('/todos/:id', authenticate, (req, res) => {
 });
 
 
-app.delete('/todos/:id', authenticate, (req, res) => {
-  var id = req.params.id;
+app.delete('/todos/:id', authenticate, async (req, res) => {
+  try {
+    const id = req.params.id;
 
-  if (!ObjectID.isValid(id)) {
-    return res.status(404).send();
-  }
+    if (!ObjectID.isValid(id)) {
+      return res.status(404).send();
+    }
 
-  Todo.findOneAndRemove({
-    _id: id,
-    _creator: req.user._id
-  }).then((todo) => {
+    const todo = await Todo.findOneAndRemove( {_id: id, _creator: req.user._id} );
     if (!todo) {
       return res.status(404).send();
     }
     res.send({todo});
-  }).catch((error) => {
+  } catch (err) {
     res.status(400).send();
-  });
+  }
 });
